refactor(client-lib): rename `fake` module to `fake_aggregator`
