+++ mithril-client/tests/cardano_transaction_proof.rs
@@ -1,6 +1,6 @@
 mod extensions;
 
-use crate::extensions::fake::{FakeAggregator, FakeCertificateVerifier};
+use crate::extensions::fake_aggregator::{FakeAggregator, FakeCertificateVerifier};
 use mithril_client::{aggregator_client::AggregatorRequest, ClientBuilder, MessageBuilder};
 
 #[tokio::test]
