+++ mithril-client/tests/certificate_get_list.rs
@@ -1,6 +1,6 @@
 mod extensions;
 
-use crate::extensions::fake::FakeAggregator;
+use crate::extensions::fake_aggregator::FakeAggregator;
 use mithril_client::{aggregator_client::AggregatorRequest, ClientBuilder};
 
 #[tokio::test]
