+++ mithril-client/tests/extensions/mod.rs
@@ -2,7 +2,7 @@
 // is a different compilation target).
 #![allow(dead_code)]
 
-pub mod fake;
+pub mod fake_aggregator;
 pub mod mock;
 mod routes;
 #[cfg(feature = "fs")]
