+++ mithril-client/tests/mithril_stake_distribution_list_get_show_verify.rs
@@ -1,6 +1,6 @@
 mod extensions;
 
-use crate::extensions::fake::{FakeAggregator, FakeCertificateVerifier};
+use crate::extensions::fake_aggregator::{FakeAggregator, FakeCertificateVerifier};
 use mithril_client::{aggregator_client::AggregatorRequest, ClientBuilder, MessageBuilder};
 
 #[tokio::test]
